updated db user specific; cost

+++ lib/pages/procedures/artificial_insemination_page.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
 class ArtificialInseminationPage extends StatefulWidget {
@@ -13,6 +14,7 @@ class _ArtificialInseminationPageState
   final TextEditingController _serialNumberController = TextEditingController();
   final TextEditingController _vetNameController = TextEditingController();
   final TextEditingController _notesController = TextEditingController();
+  final TextEditingController _costController = TextEditingController();
   String? _selectedCattle;
   String? _selectedBreed;
   String? _selectedSexed;
@@ -26,6 +28,7 @@ class _ArtificialInseminationPageState
     _serialNumberController.dispose();
     _vetNameController.dispose();
     _notesController.dispose();
+    _costController.dispose();
     super.dispose();
   }
 
@@ -81,10 +84,11 @@ class _ArtificialInseminationPageState
                       padding: const EdgeInsets.all(16.0),
                       child: Text(
                         'Artificial Insemination',
-                        style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                              color: Colors.blueAccent,
-                              fontWeight: FontWeight.bold,
-                            ),
+                        style:
+                            Theme.of(context).textTheme.headlineSmall?.copyWith(
+                                  color: Colors.blueAccent,
+                                  fontWeight: FontWeight.bold,
+                                ),
                       ),
                     ),
                     Padding(
@@ -124,10 +128,7 @@ class _ArtificialInseminationPageState
               Container(
                 width: double.infinity,
                 child: StreamBuilder<QuerySnapshot>(
-                  stream: FirebaseFirestore.instance
-                      .collection('artificial_inseminations')
-                      .limit(50)
-                      .snapshots(),
+                  stream: _getUserSpecificStream(),
                   builder: (context, snapshot) {
                     if (snapshot.connectionState == ConnectionState.waiting) {
                       return Center(child: CircularProgressIndicator());
@@ -147,6 +148,7 @@ class _ArtificialInseminationPageState
                           DataColumn(label: Text('Serial Number')),
                           DataColumn(label: Text('Vet Name')),
                           DataColumn(label: Text('Donor Breed')),
+                          DataColumn(label: Text('Cost')),
                           DataColumn(label: Text('Actions')),
                         ],
                         rows: data.docs.map<DataRow>((doc) {
@@ -156,19 +158,22 @@ class _ArtificialInseminationPageState
                             DataCell(Text(fields['serialNumber'] ?? 'N/A')),
                             DataCell(Text(fields['vetName'] ?? 'N/A')),
                             DataCell(Text(fields['breed'] ?? 'N/A')),
-                            DataCell(Row(
-                              children: [
-                                IconButton(
-                                  icon: Icon(Icons.edit, color: Colors.blue),
-                                  onPressed: () =>
-                                      _showInseminationModal(context, doc.id),
-                                ),
-                                IconButton(
-                                  icon: Icon(Icons.delete, color: Colors.red),
-                                  onPressed: () => _confirmDelete(doc.id),
-                                ),
-                              ],
-                            )),
+                            DataCell(Text(fields['cost'] ?? 'N/A')),
+                            DataCell(
+                              Row(
+                                children: [
+                                  IconButton(
+                                    icon: Icon(Icons.edit, color: Colors.blue),
+                                    onPressed: () =>
+                                        _showInseminationModal(context, doc.id),
+                                  ),
+                                  IconButton(
+                                    icon: Icon(Icons.delete, color: Colors.red),
+                                    onPressed: () => _confirmDelete(doc.id),
+                                  ),
+                                ],
+                              ),
+                            ),
                           ]);
                         }).toList(),
                       ),
@@ -181,6 +186,19 @@ class _ArtificialInseminationPageState
         ),
       ),
     );
+  }
+
+  Stream<QuerySnapshot> _getUserSpecificStream() {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) {
+      throw Exception('User not logged in');
+    }
+    return FirebaseFirestore.instance
+        .collection('users')
+        .doc(uid)
+        .collection('artificial_inseminations')
+        .limit(50)
+        .snapshots();
   }
 
   void _showInseminationModal(BuildContext context, String? docId) {
@@ -208,27 +226,44 @@ class _ArtificialInseminationPageState
                       ),
                 ),
                 SizedBox(height: 10),
-                DropdownButtonFormField<String>(
-                  decoration: InputDecoration(labelText: 'Select Cattle'),
-                  value: _selectedCattle,
-                  items: [
-                    DropdownMenuItem(value: 'Cattle1', child: Text('Cattle1')),
-                    DropdownMenuItem(value: 'Cattle2', child: Text('Cattle2')),
-                  ],
-                  onChanged: (value) {
-                    if (mounted) {
-                      setState(() {
-                        _selectedCattle = value;
-                      });
+
+                // Cattle Dropdown populated from the database
+                FutureBuilder<List<String>>(
+                  future: _fetchCattleFromDatabase(),
+                  builder: (context, snapshot) {
+                    if (snapshot.connectionState == ConnectionState.waiting) {
+                      return CircularProgressIndicator();
+                    } else if (snapshot.hasError) {
+                      return Text('Error: ${snapshot.error}');
+                    } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
+                      return Text('No cattle found');
                     }
+
+                    return DropdownButtonFormField<String>(
+                      decoration: InputDecoration(labelText: 'Select Cattle'),
+                      value: _selectedCattle,
+                      items: snapshot.data!.map((cattle) {
+                        return DropdownMenuItem(
+                            value: cattle, child: Text(cattle));
+                      }).toList(),
+                      onChanged: (value) {
+                        if (mounted) {
+                          setState(() {
+                            _selectedCattle = value;
+                          });
+                        }
+                      },
+                    );
                   },
                 ),
                 SizedBox(height: 10),
+
                 TextFormField(
                   controller: _serialNumberController,
                   decoration: InputDecoration(labelText: 'Serial Number'),
                 ),
                 SizedBox(height: 10),
+
                 TextFormField(
                   controller: _dateController,
                   decoration:
@@ -251,17 +286,40 @@ class _ArtificialInseminationPageState
                   },
                 ),
                 SizedBox(height: 10),
+
                 TextFormField(
                   controller: _vetNameController,
                   decoration: InputDecoration(labelText: 'Veterinarian Name'),
                 ),
                 SizedBox(height: 10),
+
+                // Updated Breed Dropdown
                 DropdownButtonFormField<String>(
                   decoration: InputDecoration(labelText: 'Select Breed'),
                   value: _selectedBreed,
                   items: [
-                    DropdownMenuItem(value: 'Breed1', child: Text('Breed1')),
-                    DropdownMenuItem(value: 'Breed2', child: Text('Breed2')),
+                    DropdownMenuItem(
+                        value: 'Holstein', child: Text('Holstein')),
+                    DropdownMenuItem(value: 'Angus', child: Text('Angus')),
+                    DropdownMenuItem(value: 'Jersey', child: Text('Jersey')),
+                    DropdownMenuItem(
+                        value: 'Guernsey', child: Text('Guernsey')),
+                    DropdownMenuItem(
+                        value: 'Hereford', child: Text('Hereford')),
+                    DropdownMenuItem(
+                        value: 'Simmental', child: Text('Simmental')),
+                    DropdownMenuItem(
+                        value: 'Charolais', child: Text('Charolais')),
+                    DropdownMenuItem(value: 'Brahman', child: Text('Brahman')),
+                    DropdownMenuItem(
+                        value: 'Dairy Shorthorn',
+                        child: Text('Dairy Shorthorn')),
+                    DropdownMenuItem(value: 'Dexter', child: Text('Dexter')),
+                    DropdownMenuItem(
+                        value: 'Milking Shorthorn',
+                        child: Text('Milking Shorthorn')),
+                    DropdownMenuItem(
+                        value: 'Ayrshire', child: Text('Ayrshire')),
                   ],
                   onChanged: (value) {
                     if (mounted) {
@@ -272,12 +330,20 @@ class _ArtificialInseminationPageState
                   },
                 ),
                 SizedBox(height: 10),
+
+// Updated Sexed Semen Dropdown
                 DropdownButtonFormField<String>(
-                  decoration: InputDecoration(labelText: 'Sexed Semen'),
+                  decoration: InputDecoration(labelText: 'Select Sexed Semen'),
                   value: _selectedSexed,
                   items: [
-                    DropdownMenuItem(value: 'Sexed1', child: Text('Sexed1')),
-                    DropdownMenuItem(value: 'Sexed2', child: Text('Sexed2')),
+                    DropdownMenuItem(value: 'Sexed A', child: Text('Sexed A')),
+                    DropdownMenuItem(value: 'Sexed B', child: Text('Sexed B')),
+                    DropdownMenuItem(value: 'Sexed X', child: Text('Sexed X')),
+                    DropdownMenuItem(value: 'Sexed Y', child: Text('Sexed Y')),
+                    DropdownMenuItem(
+                        value: 'Ultra Sexed', child: Text('Ultra Sexed')),
+                    DropdownMenuItem(
+                        value: 'Gendered Semen', child: Text('Gendered Semen')),
                   ],
                   onChanged: (value) {
                     if (mounted) {
@@ -287,54 +353,42 @@ class _ArtificialInseminationPageState
                     }
                   },
                 ),
+
                 SizedBox(height: 10),
+
+                // New Cost Field
+                TextFormField(
+                  controller: _costController,
+                  decoration: InputDecoration(labelText: 'Cost of AI'),
+                  keyboardType: TextInputType.number,
+                ),
+                SizedBox(height: 10),
+
                 TextFormField(
                   controller: _notesController,
-                  decoration: InputDecoration(labelText: 'Additional Notes'),
+                  decoration: InputDecoration(labelText: 'Notes'),
                 ),
                 SizedBox(height: 20),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: [
-                    ElevatedButton(
-                      onPressed: () {
-                        if (docId == null) {
-                          _submitData(context);
-                        } else {
-                          _updateData(context, docId);
-                        }
-                        Navigator.pop(context);
-                      },
-                      child: Text(
-                        docId == null ? 'Submit' : 'Update',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.blueAccent,
-                        padding:
-                            EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-                        shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(8)),
-                      ),
-                    ),
-                    SizedBox(width: 10),
-                    ElevatedButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      child: Text(
-                        'Close',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.red,
-                        padding:
-                            EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-                        shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(8)),
-                      ),
-                    ),
-                  ],
+
+                ElevatedButton(
+                  onPressed: () {
+                    if (docId == null) {
+                      _submitData(context);
+                    } else {
+                      _updateData(context, docId);
+                    }
+                    Navigator.of(context).pop();
+                  },
+                  child: Text(
+                    docId == null ? 'Submit' : 'Update',
+                    style: TextStyle(color: Colors.white),
+                  ),
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.blueAccent,
+                    padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8)),
+                  ),
                 ),
               ],
             ),
@@ -344,23 +398,31 @@ class _ArtificialInseminationPageState
     );
   }
 
-  void _resetForm() {
-    if (mounted) {
-      setState(() {
-        _dateController.clear();
-        _serialNumberController.clear();
-        _vetNameController.clear();
-        _notesController.clear();
-        _selectedCattle = null;
-        _selectedBreed = null;
-        _selectedSexed = null;
-      });
+  Future<List<String>> _fetchCattleFromDatabase() async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) {
+      throw Exception('User not logged in');
     }
+
+    final snapshot = await FirebaseFirestore.instance
+        .collection('cattle')
+        .where('userId', isEqualTo: uid)
+        .get();
+
+    return snapshot.docs.map((doc) => doc.data()['name'] as String).toList();
   }
 
   Future<void> _submitData(BuildContext context) async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) {
+      _showSnackBar('User not logged in');
+      return;
+    }
+
     try {
       await FirebaseFirestore.instance
+          .collection('users')
+          .doc(uid)
           .collection('artificial_inseminations')
           .add({
         'date': _dateController.text,
@@ -370,16 +432,26 @@ class _ArtificialInseminationPageState
         'sexed': _selectedSexed,
         'notes': _notesController.text,
         'cattle': _selectedCattle,
+        'cost': _costController.text,
       });
       _showSnackBar('AI record added successfully');
+      _resetForm();
     } catch (e) {
       _showSnackBar('Failed to add record: $e');
     }
   }
 
   Future<void> _updateData(BuildContext context, String docId) async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) {
+      _showSnackBar('User not logged in');
+      return;
+    }
+
     try {
       await FirebaseFirestore.instance
+          .collection('users')
+          .doc(uid)
           .collection('artificial_inseminations')
           .doc(docId)
           .update({
@@ -390,6 +462,7 @@ class _ArtificialInseminationPageState
         'sexed': _selectedSexed,
         'notes': _notesController.text,
         'cattle': _selectedCattle,
+        'cost': _costController.text,
       });
       _showSnackBar('AI record updated successfully');
     } catch (e) {
@@ -398,8 +471,16 @@ class _ArtificialInseminationPageState
   }
 
   Future<void> _loadDocument(String docId) async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) {
+      _showSnackBar('User not logged in');
+      return;
+    }
+
     try {
       final doc = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(uid)
           .collection('artificial_inseminations')
           .doc(docId)
           .get();
@@ -415,6 +496,7 @@ class _ArtificialInseminationPageState
             _selectedSexed = fields['sexed'];
             _serialNumberController.text = fields['serialNumber'] ?? '';
             _notesController.text = fields['notes'] ?? '';
+            _costController.text = fields['cost']?.toString() ?? '';
           });
         }
       }
@@ -443,8 +525,16 @@ class _ArtificialInseminationPageState
     );
 
     if (shouldDelete == true) {
+      final uid = FirebaseAuth.instance.currentUser?.uid;
+      if (uid == null) {
+        _showSnackBar('User not logged in');
+        return;
+      }
+
       try {
         await FirebaseFirestore.instance
+            .collection('users')
+            .doc(uid)
             .collection('artificial_inseminations')
             .doc(docId)
             .delete();
@@ -453,5 +543,19 @@ class _ArtificialInseminationPageState
         _showSnackBar('Failed to delete record: $e');
       }
     }
+  }
+
+  void _resetForm() {
+    setState(() {
+      _currentDocId = null;
+      _selectedCattle = null;
+      _selectedBreed = null;
+      _selectedSexed = null;
+      _dateController.clear();
+      _serialNumberController.clear();
+      _vetNameController.clear();
+      _notesController.clear();
+      _costController.clear();
+    });
   }
 }
